Extract bets table to its own component
